generalized the simplification step into a filtering step that accepts either a backend or a function to run on each AST, as defined in ast.filters
