+++ claripy/__init__.py
@@ -69,6 +69,7 @@ def BV(name, size, explicit_name=None): #pylint:disable=function-redefined
 from . import backend_manager as _backend_manager
 _backend_manager.backends._register_backend(_backends_module.BackendConcrete(), 'concrete', True, True)
 _backend_manager.backends._register_backend(_backends_module.BackendVSA(), 'vsa', False, False)
+_backend_manager.backends._register_backend(_backends_module.BackendSimplifier(), 'simplifier', False, False)
 
 if not os.environ.get('WORKER', False) and os.environ.get('REMOTE', False):
     try:
